refactor(themeselector.tpl.html): create .js and .css
