+++ tools/templates/libs/templates.functions.php
@@ -1,6 +1,7 @@
 <?php
 
 use YesWiki\Bazar\Service\EntryManager;
+use YesWiki\Core\Service\ThemeManager;
 
 if (!defined("WIKINI_VERSION")) {
     die("acc&egrave;s direct interdit");
@@ -395,151 +396,35 @@ function show_form_theme_selector($mode = 'selector', $formclass = '')
     }
     $listWikinames = '["'.implode('","', $listWikinames).'"]';
 
-    $selecteur = $GLOBALS['wiki']->render("@templates/themeselector.tpl.html", [
+    $wiki = $GLOBALS['wiki'];
+    $dataHtmlForPresets = array_map(function ($value) {
+        return extractDataFromPreset($value);
+    }, $wiki->config['templates'][$wiki->config['favorite_theme']]['presets'] ?? []);
+    $customCSSPresets = $wiki->services->get(ThemeManager::class)->getCustomCSSPresets() ;
+    $dataHtmlForCustomCSSPresets = array_map(function ($value) {
+        return extractDataFromPreset($value);
+    }, $customCSSPresets);
+
+    $selecteur =$wiki->render("@templates/themeselector.tpl.html", [
         'mode' => $mode,
-        'wiki' => $GLOBALS['wiki'],
+        'wiki' => $wiki,
         'id' => $id,
         'class' => $formclass,
         'bgselector' => $bgselector,
-        'themeNames' => array_keys($GLOBALS['wiki']->config['templates']),
-        'themes' => $GLOBALS['wiki']->config['templates'],
+        'themeNames' => array_keys($wiki->config['templates']),
+        'themes' => $wiki->config['templates'],
         'listWikinames' => $listWikinames,
+        'favoriteTheme' => $wiki->config['favorite_theme'] ?? null,
+        'favoriteSquelette' => $wiki->config['favorite_squelette'] ?? null,
+        'favoriteStyle' => $wiki->config['favorite_style'] ?? null,
+        'dataHtmlForPresets' => $dataHtmlForPresets,
+        'customCSSPresets' => $customCSSPresets,
+        'CUSTOM_CSS_PRESETS_PATH' => ThemeManager::CUSTOM_CSS_PRESETS_PATH,
+        'dataHtmlForCustomCSSPresets' => $dataHtmlForCustomCSSPresets,
+        'showDeleteButton' => ($wiki->UserIsAdmin()),
     ]);
 
     $js = add_templates_list_js();
-    $js .= "
-    $('.css-preset').click(function() {
-        closeNav();
-        // get data
-        var primaryColor = $(this).data('primary-color') || '#1a89a0';
-        var secondaryColor1 = $(this).data('secondary-color-1') || '#d8604c';
-        var secondaryColor2 = $(this).data('secondary-color-2') || '#d78958';
-        var neutralColor = $(this).data('neutral-color') || '#4e5056';
-        var neutralSoftColor = $(this).data('neutral-soft-color') || '#b0b1b3';
-        var neutralLightColor = $(this).data('neutral-light-color') || '#ffffff';
-        var mainTextFontsize = $(this).data('main-text-fontsize') || '17px';
-        var mainTextFontfamily = $(this).data('main-text-fontfamily') || '\'Nunito\', sans-serif';
-        var mainTitleFontfamily = $(this).data('main-title-fontfamily') || '\'Nunito\', sans-serif';
-        // set values
-        document.documentElement.style.setProperty('--primary-color', primaryColor);
-        document.documentElement.style.setProperty('--secondary-color-1', secondaryColor1);
-        document.documentElement.style.setProperty('--secondary-color-2', secondaryColor2);
-        document.documentElement.style.setProperty('--neutral-color', neutralColor);
-        document.documentElement.style.setProperty('--neutral-soft-color', neutralSoftColor);
-        document.documentElement.style.setProperty('--neutral-light-color', neutralLightColor);
-        document.documentElement.style.setProperty('--main-text-fontsize', mainTextFontsize);
-        document.documentElement.style.setProperty('--main-text-fontfamily', mainTextFontfamily);
-        document.documentElement.style.setProperty('--main-title-fontfamily', mainTitleFontfamily);
-        // set filename
-        var filename = $(this).data('key');
-        filename = filename.replace('.css','');
-        if (filename){
-            $('#preset-sidenav input.form-input[name=filename]').each(function (){
-                $(this).val(filename);
-            });
-        }
-        return false;
-    });
-    function deleteCSSPreset(elem,text,url){
-        event.preventDefault();
-        var key = $(elem).data('key');
-        var confirmResult = confirm(text);
-        if (confirmResult) {
-            $.ajax({
-                url: url,
-                success: function(data,textStatus,jqXHR){
-                    if (data.status) {
-                        console.log(key+' deleted !');
-                        $(elem).parent().remove();
-                    } else {
-                        console.log(key+' not deleted ! Message :'+JSON.stringify(data));
-                    }
-                },
-                method: 'DELETE',
-                cache: false,
-                error: function(jqXHR,textStatus,errorThrown){
-                    console.log('trying DELETE '+url+' ; but error obtained:'+textStatus);
-                },
-            });
-        }
-        // to prevent opening url
-        return false;
-    }
-    function componentToHex(c) {
-        let hex = parseInt(c).toString(16);
-        return hex.length == 1 ? '0' + hex : hex;
-    }
-    function extractFromStringWithRGB(value){
-        var res = value.match(/\s*rgb\(\s*([0-9]*)\s*,\s*([0-9]*)\s*,\s*([0-9]*)\s*\)/);
-        if (res && res.length > 3){
-            value = '#' + componentToHex(res[1]) + componentToHex(res[2]) + componentToHex(res[3]);
-        }
-        return value;
-    }
-    function getStyleValueEvenIfNotInitialized(prop){
-        var value = document.documentElement.style.getPropertyValue(prop);
-        if (!value){
-            value = getComputedStyle(document.documentElement).getPropertyValue(prop);
-        }
-        return value ;
-    }
-    function saveCSSPreset(elem,url){
-        event.preventDefault();
-        var fileName = $(elem).prev().find('input[name=filename]').val();
-        fileName = fileName.replace('.css','');
-        var fullFileName = fileName + '.css';
-        url = url + fullFileName;
-        // get values
-        var primaryColor = extractFromStringWithRGB(getStyleValueEvenIfNotInitialized('--primary-color'));
-        var secondaryColor1 = extractFromStringWithRGB(getStyleValueEvenIfNotInitialized('--secondary-color-1'));
-        var secondaryColor2 = extractFromStringWithRGB(getStyleValueEvenIfNotInitialized('--secondary-color-2'));
-        var neutralColor = extractFromStringWithRGB(getStyleValueEvenIfNotInitialized('--neutral-color'));
-        var neutralSoftColor = extractFromStringWithRGB(getStyleValueEvenIfNotInitialized('--neutral-soft-color'));
-        var neutralLightColor = extractFromStringWithRGB(getStyleValueEvenIfNotInitialized('--neutral-light-color'));
-        var mainTextFontsize = getStyleValueEvenIfNotInitialized('--main-text-fontsize');
-        var mainTextFontfamily = getStyleValueEvenIfNotInitialized('--main-text-fontfamily');
-        var mainTitleFontfamily = getStyleValueEvenIfNotInitialized('--main-title-fontfamily');
-        if (mainTextFontfamily.search(/^[A-Za-z0-9 ]*$/) != -1){
-            mainTextFontfamily = '\''+mainTextFontfamily+'\', sans-serif';
-        } else if (mainTextFontfamily.search(/^\'[A-Za-z0-9 ]*\'$/) != -1){
-            mainTextFontfamily = mainTextFontfamily+', sans-serif';
-
-        }
-        if (mainTitleFontfamily.search(/^[A-Za-z0-9 ]*$/) != -1){
-            mainTitleFontfamily = '\''+mainTitleFontfamily+'\', sans-serif';
-        } else if (mainTitleFontfamily.search(/^\'[A-Za-z0-9 ]*\'$/) != -1){
-            mainTitleFontfamily = mainTitleFontfamily+', sans-serif';
-
-        }
-        $.ajax({
-            url: url,
-            success: function(data,textStatus,jqXHR){
-                if (data.status) {
-                    console.log(fullFileName+' added !');
-                    window.location.reload();
-                } else {
-                    console.log(fullFileName+' not added !'+\"\\n\"+JSON.stringify(data));
-                }
-            },
-            method: 'POST',
-            data: {
-                'primary-color':primaryColor,
-                'secondary-color-1':secondaryColor1,
-                'secondary-color-2':secondaryColor2,
-                'neutral-color':neutralColor,
-                'neutral-soft-color':neutralSoftColor,
-                'neutral-light-color':neutralLightColor,
-                'main-text-fontsize':mainTextFontsize,
-                'main-text-fontfamily':mainTextFontfamily,
-                'main-title-fontfamily':mainTitleFontfamily,
-            },
-            cache: false,
-            error: function(jqXHR,textStatus,errorThrown){
-                console.log('trying POST '+url+' ; but error obtained:'+textStatus);
-            },
-        });
-    }
-    ";
     $GLOBALS['wiki']->addJavascript($js);
     return $selecteur;
 }
