+++ tools/templates/presentation/styles/preset-sidenav.css
@@ -1,0 +1,154 @@
+/* The side navigation menu */
+#preset-sidenav {
+height: 100%;
+/* 100% Full-height */
+width: 0;
+/* 0 width - change this with JavaScript */
+position: fixed;
+/* Stay in place */
+top: 0;
+right: 0;
+background-color: #efefef;
+border-left: 1px solid var(--neutral-soft-color);
+overflow-x: hidden;
+/* Disable horizontal scroll */
+transition: 0.5s;
+/* 0.5 second transition effect to slide in the sidenav */
+z-index: 2000;
+padding-top: 2rem;
+/* Place content under menu */
+-webkit-box-sizing: border-box;
+-moz-box-sizing: border-box;
+box-sizing: border-box;
+}
+
+#preset-form {
+margin: 0 1em;
+}
+#preset-form label {
+font-size:13px;
+font-weight: normal;
+}
+#preset-form h4 {
+font-size: 1rem;
+}
+
+/* Position and style the close button (top right corner) */
+#preset-sidenav .closebtn {
+float: right;
+font-size: 25px;
+line-height: 1.1rem;
+cursor: pointer;
+}
+
+/* Style page content - use this if you want to push the page content to the right when you open the side
+navigation */
+#yw-container {
+transition: padding-right .5s;
+}
+
+/* On smaller screens, where height is less than 450px, change the style of the sidenav (less padding and a smaller
+font size) */
+@media screen and (max-height: 450px) {
+#preset-sidenav {
+    padding-top: 15px;
+}
+
+#preset-sidenav a {
+    font-size: 18px;
+}
+}
+
+.btn-new-preset {
+margin-top: .5rem;
+}
+
+
+/* FONT PICKER */
+.font-select * {
+-webkit-box-sizing: border-box;
+box-sizing: border-box;
+}
+
+.font-select {
+font-size: 16px;
+width:100%;
+position: relative;
+display: inline-block;
+}
+
+.font-select .fs-drop {
+position: absolute;
+top: 38px;
+left: 0;
+z-index: 999;
+background: #fff;
+color: #000;
+width: 100%;
+border: 1px solid #aaa;
+border-top: 0;
+box-shadow: 0 4px 5px rgba(0,0,0,.15);
+border-radius: 0 0 4px 4px;
+}
+
+.font-select > span {
+outline: 0;
+border-radius: 0.25rem;
+border: 1px solid #ced4da;
+display: block;
+overflow: hidden;
+white-space: nowrap;
+text-overflow: ellipsis;
+height: 38px;
+line-height: 32px;
+padding: 3px 8px 3px 8px;
+color: #444;
+background: #fff url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 4 5'%3e%3cpath fill='%23303030' d='M2 0L0 2h4zm0 5L0 3h4z'/%3e%3c/svg%3e") no-repeat right 0.75rem center/8px 10px;
+-webkit-user-select: none;
+-moz-user-select: none;
+-ms-user-select: none;
+user-select: none;
+}
+
+.font-select-active > span {
+background-color: #eee;
+border-bottom-left-radius : 0;
+border-bottom-right-radius: 0;
+}
+
+.font-select .fs-results {
+max-height: 190px;
+overflow-x: hidden;
+overflow-y: auto;
+margin: 0;
+padding: 0;
+}
+
+.font-select .fs-results li {
+line-height: 80%;
+padding: 8px;
+margin: 0;
+list-style: none;
+font-size: 18px;
+white-space: nowrap;
+}
+
+.font-select .fs-results li.active {
+background-color: #3875d7;
+color: #fff;
+cursor: pointer;
+}
+
+.font-select .fs-search {
+border-bottom: 1px solid #aaa;
+padding: 4px;
+}
+
+.font-select .fs-search input {
+padding: 7px;
+width: 100%;
+border: 1px solid #aaa;
+font: 16px Helvetica, Sans-serif;
+box-shadow: inset 0 1px 3px rgba(0,0,0,.06);
+border-radius: .1875rem;
+}
